Changed the layout to have a single header.
Modified the header to be the same for logged in and anonymous pages.
Changed a bit the style of the header to be fixed and not change based on the vh and change the main div to follow resize correctly.
Enable again the returnUrl in the login to redirect to the returnUrl instead of always datasets/.
If you are already logged in it will figure as such in the header. Also if you open the login page while logged in you get redirected to datasets.
Updated all the styles removing the reference to the login-header.