feat: pass the last visited route on OIDC login (#1753)
## Description
This PR adds a service to track the last visited route, and passes it to
/oidc backend endpoint as a query parameter, so that after successful
login (i.e. when browser calls /oidc/callback), the user is redirected
to the page he started from.
Related backend PR:
https://github.com/SciCatProject/scicat-backend-next/pull/1714

## Motivation
Previously, after OIDC login, the user would always land on "/" (or
"/datasets" or some hardcoded route configured in backend). A better
user experience is to return the user to the page she was on (e.g. if
the user clicks login from /datasets/pid123/datafiles, she should come
back to this page after logging in). This PR utilizes the backend
support added in
https://github.com/SciCatProject/scicat-backend-next/pull/1714 to enable
this behavior.

## Fixes:

* None

## Changes:
- Add RotueTrackerService, to keep track of last navigated route, and
start it on Application startup.
- Modify redirectOIDC in login component to pass returnUrl (retrieved
from above service) to the backend as query param
- Don't pass redirectUrl in AuthGuard to /login component, as this is
already handled by RouteTrackerService


## Tests included
- [x] Included for each change/fix?
- [x] Passing? (Merge will not be approved unless this is checked) 

## Documentation
- [ ] swagger documentation updated \[required\]
- [ ] official documentation updated \[nice-to-have\]

### official documentation info
If you have updated the official documentation, please provide PR # and
URL of the pages where the updates are included

## Backend version
- [x] Does it require a specific version of the backend
- which version of the backend is required: v4

## Summary by Sourcery

This pull request enhances the user experience by redirecting users to
their last visited page after OIDC login. It introduces a
RouteTrackerService to keep track of the last route and passes it to the
backend as a query parameter during the OIDC authentication process. It
also adds basic exception handling in /auth-callback to ensure returnUrl
is a valid path.

New Features:
- Implements a RouteTrackerService to track the last visited route
before OIDC login.
- Passes the last visited route to the backend's /oidc endpoint as a
query parameter.

Enhancements:
- After successful OIDC login, users are now redirected to their last
visited page instead of a hardcoded route.

Tests:
- Added a unit test for the RouteTrackerService to verify that it saves
the last route navigated from.