fix: fixing related proposals error related to count (#1776)
## Description
There was an Internal server error in the related proposals count.


## Motivation
Fixing the error that was thrown in staging environment


## Fixes:
* https://jira.ess.eu/browse/SWAP-4554


## Changes:
Please provide a list of the changes implemented by this PR

* changes made


## Tests included
- [ ] Included for each change/fix?
- [ ] Passing? (Merge will not be approved unless this is checked) 

## Documentation
- [x] swagger documentation updated \[required\]
- [ ] official documentation updated \[nice-to-have\]

### official documentation info
If you have updated the official documentation, please provide PR # and
URL of the pages where the updates are included

## Backend version
- [ ] Does it require a specific version of the backend
- which version of the backend is required:

## Summary by Sourcery

Fixes an internal server error related to the related proposals count by
updating the proposalsControllerCount method to include query filters.

Bug Fixes:
- Fixes an internal server error in the related proposals count.

Enhancements:
- Updates the SciCat SDK to version 4.13.0.