test: e2e tests for proposals new dynamic material table (#1766)
## Description
This PR adds e2e tests for the new proposal table and also improve the
syncing of the default and saved table settings.


## Motivation
e2e tests for the new proposal table were missing and sync of the saved
and default settings wasn't working properly.


## Fixes:

* https://jira.ess.eu/browse/SWAP-4588
* https://jira.ess.eu/browse/SWAP-4539


## Changes:
Please provide a list of the changes implemented by this PR

* changes are mostly in the tests and some syncing logic for the tables
config.


## Tests included
- [x] Included for each change/fix?
- [x] Passing? (Merge will not be approved unless this is checked) 

## Documentation
- [ ] swagger documentation updated \[required\]
- [ ] official documentation updated \[nice-to-have\]

### official documentation info
If you have updated the official documentation, please provide PR # and
URL of the pages where the updates are included

## Backend version
- [ ] Does it require a specific version of the backend
- which version of the backend is required:

---------

Co-authored-by: Max Novelli <max.novelli@ess.eu>