feat: release-jobs backwards compatible (#1777)
## Description
This PR contains only minimal changes needed in the FE to accommodate
the new BE `release-jobs` feat.

## Motivation
https://github.com/SciCatProject/frontend/issues/1619 --> After
https://github.com/SciCatProject/scicat-backend-next/pull/1741 gets
merged in the BE, this PR will allow us to have a more detailed view of
jobs in the FE, while still being compatible with the `v3` job schema.

## Fixes:
Clicking on each table row in the Jobs dashboard is now enabled and
redirects users to a page where they can view the job details.

## Changes:
The jobs detail page has been updated so that in matches with the latest
FE changes, while still matching the job schema.

## Tests included
- [ ] Included for each change/fix?
- [ ] Passing? (Merge will not be approved unless this is checked) 

## Documentation
- [ ] swagger documentation updated \[required\]
- [ ] official documentation updated \[nice-to-have\]

### official documentation info
If you have updated the official documentation, please provide PR # and
URL of the pages where the updates are included

## Backend version
- [x] Does it require a specific version of the backend
- runs with `jobs-backwards-compatible`
- or `release-jobs` after
https://github.com/SciCatProject/scicat-backend-next/pull/1742 gets
merged

## Summary by Sourcery

Implement frontend changes to support the new backend 'release-jobs'
feature, including enabling row clicks in the Jobs dashboard to view job
details, updating the job details page, and adding JSON viewing
capabilities.

New Features:
- Enable clicking on table rows in the Jobs dashboard to redirect users
to a detailed job view page.

Enhancements:
- Update the jobs detail page to align with the latest frontend changes
and maintain compatibility with the job schema.
- Add a copy-to-clipboard feature for job IDs in the job details view.
- Integrate ngx-json-viewer for displaying JSON data in job details.