refactor: proposal-dashboard table refactor before replacing the instruments table (#1791)
## Description
Refactors the proposal dashboard component to improve data fetching and
table configuration. It enhances the component's responsiveness to route
parameters for pagination and search, and streamlines the table
initialization process.


## Motivation
There was quite much repetitive logic in the proposal-dashboard
component for the table itself that could be simplified.

## Changes:

- Improves data fetching and table configuration in the proposal
dashboard component.
- Enhances responsiveness to route parameters for pagination and search.
- Streamlines the table initialization process.
- Updates default values for table properties in the shared table
component.


## Tests included
- [ ] Included for each change/fix?
- [ ] Passing? (Merge will not be approved unless this is checked) 

## Documentation
- [ ] swagger documentation updated \[required\]
- [ ] official documentation updated \[nice-to-have\]

### official documentation info
If you have updated the official documentation, please provide PR # and
URL of the pages where the updates are included

## Backend version
- [ ] Does it require a specific version of the backend
- which version of the backend is required:

## Summary by Sourcery

Refactors the proposal dashboard component to improve data fetching and
table configuration. It enhances the component's responsiveness to route
parameters for pagination and search, and streamlines the table
initialization process.

Enhancements:
- Improves data fetching and table configuration in the proposal
dashboard component.
- Enhances responsiveness to route parameters for pagination and search.
- Streamlines the table initialization process.
- Updates default values for table properties in the shared table
component.

## Summary by Sourcery

Refactors the proposal dashboard component to improve data fetching and
table configuration. This change enhances the component's responsiveness
to route parameters for pagination and search and streamlines the table
initialization process by subscribing to route changes and dispatching
actions accordingly.

Enhancements:
- Improves data fetching and table configuration in the proposal
dashboard component.
- Enhances responsiveness to route parameters for pagination and search.
- Streamlines the table initialization process.
- Updates default values for table properties in the shared table
component.